refactor: rozdzielenie MapperHelpera oraz uprzątnięcie ArticleCreatedDto
