+++ src/main/java/org/aszjch/demoapp/aplication/MapStructConfig.java
@@ -1,12 +1,13 @@
 package org.aszjch.demoapp.aplication;
 
+import org.aszjch.demoapp.infrastructure.storage.MultipartFileToFileConverter;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.MapperConfig;
 
 @MapperConfig(
         componentModel = "spring",
         injectionStrategy = InjectionStrategy.CONSTRUCTOR,
-        uses = MapperHelper.class
+        uses = {MultipartFileToFileConverter.class, CurrentTimeProvider.class}
 )
 public interface MapStructConfig {
