+++ src/main/java/org/aszjch/demoapp/domain/article/ArticleConfig.java
@@ -2,7 +2,6 @@ package org.aszjch.demoapp.domain.article;
 
 import org.aszjch.demoapp.domain.article.port.ArticlePublisher;
 import org.aszjch.demoapp.domain.article.port.ArticleRepository;
-import org.aszjch.demoapp.domain.articlefile.ArticleFileService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -11,9 +10,8 @@ class ArticleConfig {
 
     @Bean
     ArticleService articleService(final ArticleRepository articleRepository,
-                                  final ArticlePublisher articlePublisher,
-                                  final ArticleFileService articleFileService) {
-        return new ArticleServiceImpl(articleRepository, articlePublisher, articleFileService);
+                                  final ArticlePublisher articlePublisher) {
+        return new ArticleServiceImpl(articleRepository, articlePublisher);
     }
 
 }
