+++ src/main/java/org/aszjch/demoapp/infrastructure/storage/MultipartFileToFileConverter.java
@@ -4,6 +4,7 @@ import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aszjch.demoapp.infrastructure.config.TempFilesProperties;
+import org.mapstruct.Named;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -13,6 +14,7 @@ import java.nio.file.Path;
 
 @Slf4j
 @Component
+@Named("MultipartFileToFileConverter")
 @RequiredArgsConstructor
 public class MultipartFileToFileConverter {
 
@@ -28,6 +30,7 @@ public class MultipartFileToFileConverter {
         }
     }
 
+    @Named("convert")
     public File convertMultipartFile(final MultipartFile multipartFile) throws IOException {
         final File file = Path.of(tempFilesProperties.getDir(), multipartFile.getOriginalFilename())
                 .toFile();
